requirements.txt
Pillow==8.3.1
fpdf==1.7.2
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
pdfminer.six==20221105
